requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
psycopg2-binary==2.9.9
lxml==4.9.4
qdrant-client==1.1.1
click==8.1.3
python-dotenv==0.21.0
torch>=2.0.0
sentence-transformers>=2.2.2
tqdm>=4.64.1
transformers>=4.26.0